Messup, added bones
- Made the code a lot more messy while adding stuff
+ Added bones! They're pretty much linkedlists that are a little more
complicated. Every bone can have many child bones, but every child bone
can only have one parent. The root bone is the only one that rotates,
every other bone is only 'offset'.
+ Added an angle solver and moved the radian->degree and degree->radian
functions there

BoneLink was going to be a class that linked bones together, but I
scrapped it in favor for the linkedlist approach. KeyFrame is going to
be one full 'skeleton' of bones for animation.